% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.p.fdr.R
\name{plot.p.fdr}
\alias{plot.p.fdr}
\title{FDR plotting}
\usage{
\method{plot}{p.fdr}(
  p.fdr.object,
  raw.pvalues = TRUE,
  adj.pvalues = TRUE,
  sig.line = TRUE,
  adj.sig.line = TRUE,
  threshold = NA,
  x.axis = "Rank",
  x.lim = NA,
  y.lim = c(0, 1),
  zvalues = "two.sided",
  legend.where = NA,
  title = NA,
  pch.star = 17
)
}
\arguments{
\item{p.fdr.object}{A p.fdr object that contains the list of output.}

\item{raw.pvalues}{A Boolean TRUE or FALSE value to indicate whether or not to plot the raw p-value points. Defaults to TRUE.}

\item{adj.pvalues}{A Boolean TRUE or FALSE value to indicate whether or not to plot the adjusted p-value points. Defaults to TRUE.}

\item{sig.line}{A Boolean TRUE or FALSE value to indicate whether or not to plot the raw p-value significance line. Defaults to TRUE.}

\item{adj.sig.line}{A Boolean TRUE or FALSE value to indicate whether or not to plot the adjusted significance threshold. Defaults to TRUE.}

\item{threshold}{A numeric value to determine the threshold at which we plot significance. Defaults to value used in the p.fdr.object.}

\item{x.axis}{A string variable to indicate what to plot on the x-axis. Can either be "Rank" or "Zvalues". Defaults to "Rank".}

\item{x.lim}{A numeric interval for x-axis limits.}

\item{y.lim}{A numeric interval for y-axis limits. Defaults to c(0,1).}

\item{zvalues}{A numeric vector of z-values to be used in pi0 estimation or a string with options "two.sided", "greater" or "less". Defaults to "two.sided".}

\item{legend.where}{A string "bottomright", "bottomleft", "topleft", "topright". Defaults to "topleft" is x.axis="Rank" and "topright" if x.axis="Zvalues".}

\item{title}{A string variable for the title of the plot.}

\item{pch.star}{A plotting ‘character’, or symbol to use for the adjusted p-value points. This can either be a single character or an integer code for one of a set of graphics symbols. Defaults to 17.}
}
\description{
This function creates a plot using a p.fdr.object.
}
\details{
We run into errors or warnings when
}
\examples{

# Example 1

sim.data.p = c(runif(80),runif(20, min=0, max=0.01))
fdr.output = p.fdr(pvalues=sim.data.p, sort.results = TRUE)

plot(fdr.output)
plot(fdr.output, x.axis="Zvalues")

}
\references{

}
\seealso{
\code{\link{summary.p.fdr}, \link{p.fdr}, \link{get.pi0}}
}
