% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.p.fdr.R
\name{print.summary.p.fdr}
\alias{print.summary.p.fdr}
\title{Print the summary of p.fdr.object}
\usage{
\method{print}{summary.p.fdr}(summary.p.fdr.object, digits = 3)
}
\arguments{
\item{summary.p.fdr.object}{A list of output from the summary.p.fdr function.}

\item{digits}{A numeric value for the number of desired digits in the summary output. Defaults to 3.}
}
\description{
This function prints the summary a p.fdr.object.
}
\details{
We run into errors or warnings when
}
\examples{

# Example 1
pi0 = 0.8
pi1 = 1-pi0
n = 10
n.0 = ceiling(n*pi0)
n.1 = n-n.0

sim.data = c(rnorm(n.1,5,1),rnorm(n.0,0,1))
sim.data.p = 2*pnorm(-abs(sim.data))

fdr.output = p.fdr(pvalues=sim.data.p, adjust.method="BH")

summary(fdr.output)

}
\references{
R Journal 2020?
}
\seealso{
\code{\link{plot.p.fdr}, \link{p.fdr}, \link{get.pi0}}
}
\keyword{FDR,}
\keyword{adjusted}
\keyword{null}
\keyword{p-values,}
\keyword{proportion}
\keyword{summary,}
